+++ App.js
@@ -1,38 +1,20 @@
-import React from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
-import Character from './src/pages/Character';
-import Start from './src/pages/start';
-import Questions from './src/pages/question';
-import Login from './src/pages/login';
-
-const Stack = createStackNavigator();
+import { StatusBar } from 'expo-status-bar';
+import { StyleSheet, Text, View } from 'react-native';
 
 export default function App() {
   return (
-    <NavigationContainer>
-      <Stack.Navigator>
-        <Stack.Screen
-          name="Start"
-          options={{ headerShown: false }}
-          component={Start}
-        />
-        <Stack.Screen
-          name="Choice"
-          options={{ headerShown: false }}
-          component={Character}
-        />
-        <Stack.Screen
-          name="Questions"
-          options={{ headerShown: false }}
-          component={Questions}
-        />
-        <Stack.Screen
-          name="Login"
-          options={{ headerShown: false }}
-          component={Login}
-        />
-      </Stack.Navigator>
-    </NavigationContainer>
+    <View style={styles.container}>
+      <Text>Open up App.js to start working on your app!</Text>
+      <StatusBar style="auto" />
+    </View>
   );
 }
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
