--- index.js
@@ -1,4 +1,0 @@
-import { registerRootComponent } from "expo";
-import App from "./App";
-
-registerRootComponent(App);
