+++ package.json
@@ -9,20 +9,10 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@react-navigation/native": "^6.1.18",
-    "@react-navigation/native-stack": "^6.11.0",
-    "@react-navigation/stack": "^6.4.1",
     "expo": "~51.0.28",
     "expo-status-bar": "~1.12.1",
     "react": "18.2.0",
-    "react-native": "0.74.5",
-    "react-native-gesture-handler": "^2.20.2",
-    "react-native-image-slider-box": "^2.0.7",
-    "react-native-responsive-screen": "^1.4.2",
-    "react-native-safe-area-context": "^4.14.0",
-    "react-native-screens": "^3.35.0",
-    "react-native-snap-carousel": "^3.9.1",
-    "react-native-swiper": "^1.6.0"
+    "react-native": "0.74.5"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
