--- src/components/bottomBar.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export default function BottomBar() {
-  return (
-    <View>
-      <Text>하단바</Text>
-    </View>
-  );
-}
