--- src/pages/Character.jsx
@@ -1,12 +1,0 @@
-import { View, Text } from 'react-native';
-import React from 'react';
-
-export default function Character() {
-  return (
-    <View>
-      <View>
-        <Text>캐릭터</Text>
-      </View>
-    </View>
-  );
-}
