--- src/pages/login.jsx
@@ -1,9 +1,0 @@
-import { View, Text } from 'react-native';
-
-export default function Login() {
-  return (
-    <View>
-      <Text>로그인</Text>
-    </View>
-  );
-}
