--- src/pages/question.jsx
@@ -1,9 +1,0 @@
-import { View, Text } from 'react-native';
-
-export default function Questions() {
-  return (
-    <View>
-      <Text>질문</Text>
-    </View>
-  );
-}
